Add supportcase_discard_mail_aliases setting
This adds a new setting that accepts known aliases of configured
inboxes that should be discarded when prefilling To and CC fields
in the mail editor. This is necessary so we don't send emails to
ourselves if an inbound email is sent to an alias rather than the
main address that is configured in the corresponding MailutilsSetting
FromEmailAddress.